Support multiple networks for vSphere Micro
This commit adds a new optional section in the micro deploy manifest called
`deployment_network`. It will be a (restricted) app network. The `network`
section is expected to contain a more exposed network that is reachable from
the app networks.

Use deployment_network ip by default and a vip (if exists) as a fallback when
updating agent service addresses

Signed-off-by: John Foley <jfoley@pivotallabs.com>