Don't rely on 'df' inside warden
There was a pre-existing guard in a test in with_release_stemcell_spec
to not run 'df' within warden as it may return "df: no file systems
processed".  This seems to be a known issue.

This change adds a guard around another call to 'df', done in a before
block for the aforementioned test.

Signed-off-by: Andrea Law <alaw@pivotallabs.com>